A fix for geting past off having 3 diffrent Auction House as intended. But there where some things that i neede to do change for it to be able to do that so there are a new option in worldserver.conf and its called AuctionHouseBot.OverrideSide If you set that one to 1 it will works as it should even if you have ally and horde working togehter.
modified:   doc/AuctionHouseBot.txt
	modified:   src/server/game/AuctionHouse/AuctionHouseBot/AuctionHouseBot.cpp
	modified:   src/server/game/AuctionHouse/AuctionHouseBot/AuctionHouseBot.h
	modified:   src/server/game/AuctionHouse/AuctionHouseMgr.cpp
	modified:   src/server/game/AuctionHouse/AuctionHouseMgr.h

+++ src/server/game/AuctionHouse/AuctionHouseMgr.h
@@ -59,7 +59,7 @@ struct AuctionEntry
     uint32 auctioneer;                                      // creature low guid
     uint32 item_guidlow;
     uint32 item_template;
-    uint64 owner;
+    uint32 owner;
     uint32 startbid;                                        //maybe useless
     uint32 bid;
     uint32 buyout;
